Pull `FileServiceConfig` into `BaseFileService` as an inner class.
