+++ src/sys-config/export/SaveConfig.js
@@ -4,12 +4,10 @@
 import { BaseConfig } from '@this/compote';
 import { MustBe } from '@this/typey';
 
-import { FileServiceConfig } from '#x/FileServiceConfig';
-
 
 /**
  * Configuration representation for file preservation, used in configuring some
- * file-writing services. (See {@link FileServiceConfig}.) See
+ * file-writing services. See
  * `doc/configuration` for configuration object details.
  */
 export class SaveConfig extends BaseConfig {
