+++ src/sys-config/index.js
@@ -1,7 +1,6 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '#x/FileServiceConfig';
 export * from '#x/RotateConfig';
 export * from '#x/SaveConfig';
 export * from '#x/ServiceConfig';
